requirements.txt
discord.py[voice]
ffmpeg
intent
